[CCUBE-1433][SR] Made a file to keep text-based field schema generation in.
